PEPPER-833 Moved non-DB tests to KitRequestShippingTest.java and changed where I check for error in KitTypeImpl.getKitTypes
