+++ pepper-apis/dsm-core/src/test/java/org/broadinstitute/dsm/kits/KitDisplayNameTest.java
@@ -2,7 +2,6 @@ package org.broadinstitute.dsm.kits;
 
 import java.util.List;
 
-import com.google.gson.Gson;
 import org.broadinstitute.dsm.DbTxnBaseTest;
 import org.broadinstitute.dsm.db.DDPInstance;
 import org.broadinstitute.dsm.db.KitRequestShipping;
@@ -16,11 +15,8 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 
 public class KitDisplayNameTest extends DbTxnBaseTest {
-    private static String guid = "TEST_GUID";
     private static String BLOOD_RNA_KIT_TYPE_NAME = "BLOOD";
     private static String BLOOD_RNA_KIT_TYPE_DISPLAY_NAME = "BLOOD and RNA";
-    static String[] subkitSettingsIds = null;
-
     private static TestKitUtil
             testKitUtil = new TestKitUtil("test_instance", "test_instance_guid", "some_prefix", "test_group", BLOOD_RNA_KIT_TYPE_NAME);
 
@@ -32,78 +28,20 @@ public class KitDisplayNameTest extends DbTxnBaseTest {
     @AfterClass
     public static void cleanUpAfterClass() {
         testKitUtil.deleteKitsArray();
-        if (subkitSettingsIds != null){
-            testKitUtil.deleteSubKitSettings(subkitSettingsIds);
-        }
         testKitUtil.deleteInstanceAndSettings();
-
     }
 
     @Test
-    public void testKitWithDisplayNAme(){
+    public void testKitWithDisplayName(){
         NonPepperKitCreationService nonPepperKitCreationService = new NonPepperKitCreationService();
-        DDPInstance ddpInstance = DDPInstance.getDDPInstanceWithRole("test_instance",DBConstants.JUNIPER_STUDY_INSTANCE_ROLE);
-        JuniperKitRequest juniperTestKit = generateKitRequestJson();
-        testKitUtil.createNonPepperTestKit(juniperTestKit, nonPepperKitCreationService, ddpInstance);
+        DDPInstance ddpInstance = DDPInstance.getDDPInstanceWithRole("test_instance", DBConstants.JUNIPER_STUDY_INSTANCE_ROLE);
+        JuniperKitRequest juniperTestKit = TestKitUtil.generateKitRequestJson();
+        TestKitUtil.createNonPepperTestKit(juniperTestKit, nonPepperKitCreationService, ddpInstance);
         List<KitRequestShipping> kits = KitRequestShipping.getKitRequestsByRealm("test_instance", "overview", BLOOD_RNA_KIT_TYPE_NAME);
         Assert.assertEquals(1, kits.size());
-        Assert.assertEquals("BLOOD", kits.get(0).getKitTypeName());
-        Assert.assertEquals("BLOOD and RNA", kits.get(0).getDisplayName());
+        Assert.assertEquals(BLOOD_RNA_KIT_TYPE_NAME, kits.get(0).getKitTypeName());
+        Assert.assertEquals(BLOOD_RNA_KIT_TYPE_DISPLAY_NAME, kits.get(0).getDisplayName());
         Assert.assertNotEquals(kits.get(0).getKitTypeName(), kits.get(0).getDisplayName());
 
-    }
-
-    @Test
-    public void testEmptyDisplayName() {
-
-        // test that not having a display name means the display name is set as the kit type name
-        KitRequestShipping kitWithoutDisplayName = new KitRequestShipping(guid, "TestProject_2", null, "FAKE_DSM_LABEL_UID", "study", BLOOD_RNA_KIT_TYPE_NAME, 1L, 1L,
-                "https://easypost-files.s3-us-west-2.amazonaws"
-                        + ".com/files/postage_label/20200214/8240f1b66535494a82b1ec0d566c3f0f.png", "", "794685038506",
-                "9405536897846100551129", "https://track.easypost.com/djE6dHJrXzY4NGJmYzU3ZjM5OTQ1Zjg5MjEzOGRmMWVmMjI1NWZl",
-                null, 12L, false, "", 12L, null, 12L, "so what", "mf_testLabel", false, "shp_f470591c3fb441a68dbb9b76ecf3bb3d",
-                12L, null, "44445", false, "NOT FOUND", null, null, null, null, 0L, false, "STANDALONE", null, null, null,
-                null, null, null, null, null, null, null);
-        Assert.assertEquals(kitWithoutDisplayName.getDisplayName(), BLOOD_RNA_KIT_TYPE_NAME);
-        Assert.assertEquals(kitWithoutDisplayName.getKitTypeName(), BLOOD_RNA_KIT_TYPE_NAME);
-
-
-    }
-
-    @Test
-    public void testNotEmptyDisplayName() {
-
-        // test that not having a display name means the display name is set as the kit type name
-        KitRequestShipping kitWithDisplayName = new KitRequestShipping(guid, "TestProject_2", null, "FAKE_DSM_LABEL_UID", "study", BLOOD_RNA_KIT_TYPE_NAME, 1L, 1L,
-                "https://easypost-files.s3-us-west-2.amazonaws"
-                        + ".com/files/postage_label/20200214/8240f1b66535494a82b1ec0d566c3f0f.png", "", "794685038506",
-                "9405536897846100551129", "https://track.easypost.com/djE6dHJrXzY4NGJmYzU3ZjM5OTQ1Zjg5MjEzOGRmMWVmMjI1NWZl",
-                null, 12L, false, "", 12L, null, 12L, "so what", "mf_testLabel", false, "shp_f470591c3fb441a68dbb9b76ecf3bb3d",
-                12L, null, "44445", false, "NOT FOUND", null, null, null, null, 0L, false, "STANDALONE", null, null, null,
-                null, null, null, null, null, null, BLOOD_RNA_KIT_TYPE_DISPLAY_NAME);
-        Assert.assertEquals(kitWithDisplayName.getDisplayName(), BLOOD_RNA_KIT_TYPE_DISPLAY_NAME);
-        Assert.assertEquals(kitWithDisplayName.getKitTypeName(), BLOOD_RNA_KIT_TYPE_NAME);
-
-
-    }
-
-    private static JuniperKitRequest generateKitRequestJson(){
-            String participantId = "TEST_PARTICIPANT";
-
-            String json = "{ \"firstName\":\"P\","
-                    + "\"lastName\":\"T\","
-                    + "\"street1\":\"415 Main st\","
-                    + "\"street2\":null,"
-                    + "\"city\":\"Cambridge\","
-                    + "\"state\":\"MA\","
-                    + "\"postalCode\":\"02142\","
-                    + "\"country\":\"USA\","
-                    + "\"phoneNumber\":\" 111 - 222 - 3344\","
-                    + "\"juniperKitId\":\"kitId_\","
-                    + "\"juniperParticipantID\":\"" + participantId  + "\","
-                    + "\"skipAddressValidation\":false,"
-                    + "\"juniperStudyID\":\"Juniper-test-guid\"}";
-
-            return new Gson().fromJson(json, JuniperKitRequest.class);
     }
 }
